- compare on TestInStream constructor fixed
git-svn-id: svn://svn.code.sf.net/p/cdcat/svn/trunk@637 23e97676-e3e9-41ff-bed1-4a4288a26c68

+++ src/dbase.cpp
@@ -100,7 +100,7 @@ class TestInStream : public C7ZipInStream {
 
 				unsigned int pos = m_strFileName.find_last_of ( "." );
 
-				if ( pos != m_strFileName.npos ) {
+				if ( pos != (unsigned int)(m_strFileName.npos) ) {
 					QString ext = QString::fromStdString ( m_strFileName).split('.').last();
 					m_strFileExt = ext.toStdWString();
 				}
